+++ anxcloud/provider.go
@@ -51,7 +51,6 @@ func Provider() *schema.Provider {
 			"anxcloud_vlans":                 dataSourceVLANs(),
 			"anxcloud_tags":                  dataSourceTags(),
 			"anxcloud_cpu_performance_types": dataSourceCPUPerformanceTypes(),
-			"anxcloud_vsphere_locations":     dataSourceVSphereLocations(),
 			"anxcloud_dns_records":           dataSourceDNSRecords(),
 			"anxcloud_dns_zones":             datasourceDNSZones(),
 		},