+++ anxcloud/strings.go
@@ -4,8 +4,5 @@ const (
 	identifierDescription = "Identifier of the API resource."
 
 	// Lists
-	// SYSENG-1273 will remove manual pagination
-	listPageIndexDescription = "The page of the list."
-	listLimitDescription     = "The limit of the list."
-	listQueryDescription     = "An optional string allowing to search through entities."
+	listQueryDescription = "An optional string allowing to search through entities."
 )