+++ anxcloud/struct_locations.go
@@ -4,7 +4,6 @@ import (
 	corelocation "go.anx.io/go-anxcloud/pkg/core/location"
 	"go.anx.io/go-anxcloud/pkg/ipam/prefix"
 	"go.anx.io/go-anxcloud/pkg/vlan"
-	"go.anx.io/go-anxcloud/pkg/vsphere/provisioning/location"
 )
 
 // expanders
@@ -80,29 +79,4 @@ func flattenVLANLocations(in []vlan.Location) []interface{} {
 	}
 
 	return att
-}
-
-func flattenVSphereLocations(in []location.Location) []interface{} {
-
-	att := []interface{}{}
-	if len(in) < 1 {
-		return att
-	}
-
-	for _, d := range in {
-		m := map[string]interface{}{}
-
-		m["code"] = d.Code
-		m["country"] = d.Country
-		m["identifier"] = d.ID
-		m["lat"] = d.Latitude
-		m["lon"] = d.Longitude
-		m["name"] = d.Name
-		m["country_name"] = d.CountryName
-
-		att = append(att, m)
-	}
-
-	return att
-
 }