+++ examples/data-sources/anxcloud_vsphere_locations/data-source.tf
@@ -1,4 +1,0 @@
-data "anxcloud_vsphere_locations" "example" {
-  page  = 1
-  limit = 50
-}