+++ CHANGELOG.md
@@ -23,6 +23,7 @@ If the change isn't user-facing but still relevant enough for a changelog entry,
   - anxcloud_ip_addresses
   - anxcloud_tags
   - anxcloud_vlans
+* data-source/anxcloud_vsphere_locations: removed (previously marked deprecated; you can just use `anxcloud_core_location` everywhere this was needed before) (#104, @marioreggiori)
   
 ### Fixed
 * resource/anxcloud_virtual_server: tags changed outside of terraform will now get reverted back to terraform config on apply (#101, @marioreggiori)